feat: add more error response
style: format solution

+++ CwkSocial.Application/UserProfiles/CommandHandlers/DeleteUserProfileCommandHandler.cs
@@ -1,28 +1,36 @@
-﻿using CwkSocial.Application.UserProfiles.Commands;
+﻿using CwkSocial.Application.Enums;
+using CwkSocial.Application.Models;
+using CwkSocial.Application.UserProfiles.Commands;
 using CwkSocial.Dal;
+using CwkSocial.Domain.Aggregates.UserProfileAggregate;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace CwkSocial.Application.UserProfiles.CommandHandlers {
 
-    internal class DeleteUserProfileCommandHandler : IRequestHandler<DeleteUserProfile> {
+    internal class DeleteUserProfileCommandHandler : IRequestHandler<DeleteUserProfile, OperationResult<UserProfile>> {
         private readonly DataContext _ctx;
 
         public DeleteUserProfileCommandHandler(DataContext ctx) {
             _ctx = ctx;
         }
 
-        public async Task<Unit> Handle(DeleteUserProfile request, CancellationToken cancellationToken) {
+        public async Task<OperationResult<UserProfile>> Handle(DeleteUserProfile request, CancellationToken cancellationToken) {
+            var result = new OperationResult<UserProfile>();
             var userProfile = await _ctx.UserProfiles.FirstOrDefaultAsync(up => up.UserProfileId == request.UserProfileId);
+
+            if (userProfile is null) {
+                result.IsError = true;
+                var error = new Error { Code = ErrorCode.NotFound, Message = $"No UserProfile found with ID {request.UserProfileId}" };
+                result.Errors.Add(error);
+                return result;
+            }
 
             _ctx.UserProfiles.Remove(userProfile);
             await _ctx.SaveChangesAsync();
-            return new Unit();
+
+            result.Payload = userProfile;
+            return result;
         }
     }
 }
